feat(fin-021): refactor plus button visibility logic and update imports for user nav store
