+++ src/app/(profile)/profile/budget/(regular-operations)/regular-operations/add/page.tsx
@@ -2,7 +2,7 @@
 
 import { RegularPaymentForm } from '@frontend/features/regular-incomes-expenses/card-creation-form/regular-card-form';
 import { useRouter } from 'next/navigation';
-import { useHidePlusButton } from '@frontend/components/mobile-navbar/use-hide-plus-button';
+import { useHidePlusButton } from '@frontend/widgets/profile-mobile-navbar/use-hide-plus-button';
 
 export default function CreateCardForm() {
   const router = useRouter();
