+++ src/app/(profile)/profile/budget/(regular-operations)/regular-operations/edit/[id]/page.tsx
@@ -5,7 +5,7 @@ import { useRegularTransactions } from '@frontend/features/regular-incomes-expen
 import { useRouter } from 'next/navigation';
 import { RegularPaymentForm } from '@frontend/features/regular-incomes-expenses/card-creation-form/regular-card-form';
 import { FinFormScreenHandler } from '@frontend/components/form-screen-handler/fin-form-screen-handler';
-import { useHidePlusButton } from '@frontend/components/mobile-navbar/use-hide-plus-button';
+import { useHidePlusButton } from '@frontend/widgets/profile-mobile-navbar/use-hide-plus-button';
 
 export default function EditCardForm({ params }: IdPromiseParamsModel) {
   useHidePlusButton();
